requirements.txt
pyuseragents~=1.0.5
names~=0.3.0
loguru~=0.7.2
solders~=0.21.0
mnemonic~=0.21
PyYAML~=6.0.1
pydantic~=2.6.4
art~=6.1
orjson~=3.9.15
httpx~=0.27.0
tqdm~=4.66.2
curl_cffi
httpx
noble_tls
aiofiles